Extract csv download for windows logic to custom hook
